+++ plugin/src/test/kotlin/org/jlleitschuh/gradle/ktlint/GitHookTasksTest.kt
@@ -157,8 +157,10 @@ class GitHookTasksTest : AbstractPluginTest() {
         build(":$INSTALL_GIT_HOOK_FORMAT_TASK").run {
             assertThat(task(":$INSTALL_GIT_HOOK_FORMAT_TASK")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
             val hookText = gitDir.preCommitGitHook().readText()
-            assertThat(hookText).contains("git stash push")
-            assertThat(hookText).contains("git stash pop")
+            assertThat(hookText).contains("git diff --color=never > \$diff")
+            assertThat(hookText).contains("git apply -R \$diff")
+            assertThat(hookText).contains("git apply --ignore-whitespace \$diff")
+            assertThat(hookText).contains("rm \$diff")
         }
     }