+++ plugin/src/test/kotlin/org/jlleitschuh/gradle/ktlint/testdsl/testAnnotations.kt
@@ -13,7 +13,8 @@ object TestVersions {
     const val minSupportedGradleVersion = KtlintBasePlugin.LOWEST_SUPPORTED_GRADLE_VERSION
     const val maxSupportedGradleVersion = "7.1.1"
     const val pluginVersion = "10.2.0-SNAPSHOT"
-    const val kotlinPluginVersion = "1.4.32"
+    const val minSupportedKotlinPluginVersion = "1.4.32"
+    const val maxSupportedKotlinPluginVersion = "1.5.21"
 }
 
 @Target(AnnotationTarget.FUNCTION, AnnotationTarget.CLASS)